requirements.txt
pytest
requests
python-dotenv
allure-pytest